Update database schema to support "TaglineTags"
Uhh, it would probably be good if I had a better (more readable) way of
storing the schema, but for now this dummy database sql is it

So I created a new table to store all of the tags that can be displayed
in the tagline. And some of them will display the tag label itself, and
some have an alternate display, because grammar. And some are designated
long tags and some are short tags